supercut {wbggeo}	R Documentation

(Better) Convert Numeric to Factor

Description

Like cut, supercut divides the range of x into intervals and codes the values in x according to the interval in which they fall. Unlike cut, it understands mathematical interval notation so you can be very precise with how the intervals are constructed.

Usage

supercut(x, intervals, labels = NULL, ordered_result = T)

Arguments

x
a numeric vector which is to be converted to a factor by supercutting

intervals
a vector of string intervals, see Details and Examples

labels
labels (if not given names(intervals) will be used if set)

ordered_result
logical: should the result be an ordered factor?

Details

The intervals are strings being either a single number (e.g. "10"), which will be assigned to only that value, or an interval (e.g. "[10,20]"), which will be assigned to any value in that interval. Open ("(1,5)"), closed ("[1,5]") and mixed ("[1,5)") intervals are supported. To create an interval which is open below, use e.g. "(-Inf, 5]" and for an interval which is open above use e.g. "(20, Inf)".

If a value in x doesn't match any interval, a warning will be issued.

If intervals overlap behaviour is unspecified (don't do this!)

See Also

cut

Examples

supercut(mtcars$hp, c(
  "less than 110 hp" = "(-Inf,110)",
  "exactly 110 hp"   = "110",
  "110 - 200 hp"     = "(110,200)",
  "200 hp and over"  = "[200, Inf)"
))

[Package wbggeo version 0.1.0 Index]